Extract ID pre- and post- persistence processing into IdTransformer util.
